requirements.txt
greenlet==3.2.4
numpy==2.3.3
pandas==2.3.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
tzdata==2025.2
xlwings==0.33.15
